# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing security rules, assets, and SECURITY_EXCEPTION rows.",
        "title": "Security Rules API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/de/securities/rules",
    "paths": {
...
        "/v1/api/executeRules": {
            "get": {
                "description": "Runs every rule for the given process_type against the asset and inserts any returned rows as security exceptions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rules"
                ],
                "summary": "Execute rules for an asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Process type",
                        "name": "process_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "asset_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bloomberg global ID",
                        "name": "id_bb_global",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "rules executed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "process_type and asset_id are required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to execute rules",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/api/executeSecurityRules": {
            "get": {
                "description": "Runs every rule for the given process_type against the asset and writes results into the legacy SECURITY_EXCEPTION table. Kept alongside /executeRules (which writes to the slim EXCEPTION table) during the migration.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rules"
                ],
                "summary": "Execute rules for an asset (legacy SECURITY_EXCEPTION flow)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Process type",
                        "name": "process_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "asset_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bloomberg global ID",
                        "name": "id_bb_global",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "rules executed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "process_type and asset_id are required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to execute security rules",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/api/getAssets": {
            "get": {
                "description": "Returns the deduplicated set of assets, optionally filtered by exception type CODE and severity (CATEGORY_TYPE.CODE).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assets"
                ],
                "summary": "List assets with exception summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EXCEPTION_TYPE.CODE filter",
                        "name": "exception_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "CATEGORY_TYPE.CODE filter",
                        "name": "severity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SEVERITY_TYPE.CODE filter",
                        "name": "priority",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RULE_CATALOG.NAME filter",
                        "name": "rule_catalog",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RULE.RULE_NAME filter",
                        "name": "rule_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "EXCEPTION_STATUS.CODE filter",
                        "name": "exception_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "DM_USER.USER filter",
                        "name": "assign_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/securityrules_security-rules_internal_app_models.Asset"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to query assets",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/api/getDMUsers": {
            "get": {
                "description": "Returns DM_USER.USER values ordered by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List DM users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to query users",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/api/getExceptionStatus": {
            "get": {
                "description": "Returns EXCEPTION_STATUS.CODE values ordered by SORT_ORDER.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exception-status"
                ],
                "summary": "List exception status codes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to query exception status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/api/getExceptionTypes": {
            "get": {
                "description": "Returns EXCEPTION_TYPE.NAME values ordered by SORT_ORDER.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exception-types"
                ],
                "summary": "List exception type names",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to query exception types",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/api/getPriorityTypes": {
            "get": {
                "description": "Returns EXCEPTION_PRIORITY_TYPE.NAME values ordered by SORT_ORDER.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "priority-types"
                ],
                "summary": "List priority names",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to query priority types",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "securityrules_security-rules_internal_app_models.Asset": {
            "type": "object",
            "properties": {
                "all_complete": {
                    "type": "boolean"
                },
                "asset_id": {
                    "type": "string"
                },
                "assign_to": {
                    "type": "string"
                },
                "bbg_last_refresh": {
                    "type": "string"
                },
                "exception_count": {
                    "type": "integer"
                },
                "exception_date": {
                    "type": "string"
                },
                "figi": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "security_description": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "trader": {
                    "type": "string"
                },
                "trading_team": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}